services/nema-core/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1

# Temporal Workflow
temporalio==1.4.0

# AWS Services
boto3==1.34.0
botocore==1.34.0

# Authentication
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
jwcrypto==1.5.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.0

# Data Validation
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# JSON/Data Processing  
orjson==3.9.10
python-dateutil==2.8.2

# Vector Operations
numpy==1.24.4
pgvector==0.2.4

# Monitoring/Logging
structlog==23.2.0
prometheus-client==0.19.0

# Environment/Config
python-dotenv==1.0.0

# Development/Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# File Processing
python-magic==0.4.27
Pillow==10.1.0

# Utilities
click==8.1.7
rich==13.7.0
typer==0.9.0